adding items to cart saves to database, clearing cart clears collection on database
