+++ src/components/cart/cart.component.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import {
   CartBottomContainer,
   CartButton,
@@ -15,7 +15,7 @@ import CartItem from '../cart-item/cart-item.component';
 import { selectCartItems } from '../../redux/cart/cart.selector';
 import { useNavigate } from 'react-router-dom';
 import { auth, app2 } from '../../api/firebaseConfig';
-import { doc, deleteDoc } from 'firebase/firestore';
+import { doc, deleteDoc, getDoc } from 'firebase/firestore';
 import { getFirestore } from 'firebase/firestore';
 
 const Cart = () => {
@@ -24,6 +24,26 @@ const Cart = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const database = getFirestore(app2);
+
+  useEffect(() => {
+    if (auth && auth.currentUser) {
+      (async () => {
+        const docRef = doc(database, 'Cart', auth.currentUser.email);
+        const docSnap = await getDoc(docRef);
+
+        if (docSnap.exists()) {
+          const data = docSnap.data();
+          const dataArray = Object.keys(data).map((key) => {
+            return { ...data[key] };
+          });
+          dispatch(updateCartItems(dataArray));
+        } else {
+          console.log('Document Not Found:');
+        }
+      })();
+    }
+    // eslint-disable-next-line
+  }, [database, dispatch, auth?.currentUser?.email]);
 
   const clearCartInDB = async () => {
     await deleteDoc(doc(database, 'Cart', auth.currentUser.email));
@@ -78,6 +98,7 @@ const Cart = () => {
 
           <div>
             <ClearCartButton
+              title='Clear the cart, to empty it permanently'
               colors={themeColor}
               onClick={(e) => {
                 e.preventDefault();
